Make temp float

+++ lib/features/app/domain/entities/forecast_weather_entity.dart
@@ -81,10 +81,10 @@ class MainData {
 
   factory MainData.fromJson(Map<String, dynamic> json) {
     return MainData(
-      temp: json['temp'],
+      temp: json['temp'].toDouble(),
       feelsLike: json['feels_like'].toDouble(),
-      tempMin: json['temp_min'],
-      tempMax: json['temp_max'],
+      tempMin: json['temp_min'].toDouble(),
+      tempMax: json['temp_max'].toDouble(),
       pressure: json['pressure'],
       seaLevel: json['sea_level'],
       grndLevel: json['grnd_level'],
